Drop match() method from annotation query classess
Make *Query class inherit directly from MatchExpression, drop
*MatchExpression class and use `__init__()` instead of `match()`.

Also rename files once the name has changed
