+++ lightly_studio/src/lightly_studio/core/dataset_query/__init__.py
@@ -1,9 +1,9 @@
 from .boolean_expression import AND, NOT, OR
-from .classification_expression import ClassificationField, ClassificationQuery
+from .classification_query import ClassificationField, ClassificationQuery
 from .dataset_query import DatasetQuery
 from .evaluation_metric_expression import EvaluationMetricField
 from .image_sample_field import ImageSampleField
-from .object_detection_expression import ObjectDetectionField, ObjectDetectionQuery
+from .object_detection_query import ObjectDetectionField, ObjectDetectionQuery
 from .order_by import (
     OrderByEvaluationMetricField,
     OrderByExpression,
@@ -11,7 +11,7 @@ from .order_by import (
     OrderByMetadataField,
 )
 from .sample_evaluation_query import SampleEvaluationQuery
-from .segmentation_mask_expression import SegmentationMaskField, SegmentationMaskQuery
+from .segmentation_mask_query import SegmentationMaskField, SegmentationMaskQuery
 from .video_sample_field import VideoSampleField
 
 __all__ = [
